Migrate to .NET5 and add code analyzers to all projects (#11)
* migrate to net5

* add analyzers and fix code

* fix blocking collection tests
