+++ ConcurrentPriorityQueue/ConcurrentPriorityByIntegerQueue.cs
@@ -2,11 +2,13 @@
 
 namespace ConcurrentPriorityQueue
 {
-	public class ConcurrentPriorityByIntegerQueue<T> : ConcurrentPriorityQueue<T, int>
-		where T : IHavePriority<int>
-	{
-		public ConcurrentPriorityByIntegerQueue() : base() { }
+    public class ConcurrentPriorityByIntegerQueue<T> : ConcurrentPriorityQueue<T, int>
+        where T : IHavePriority<int>
+    {
+        public ConcurrentPriorityByIntegerQueue()
+            : base() { }
 
-		public ConcurrentPriorityByIntegerQueue(int capacity) : base(capacity) { }
-	}
+        public ConcurrentPriorityByIntegerQueue(int capacity)
+            : base(capacity) { }
+    }
 }
