+++ ConcurrentPriorityQueue/Core/ConcurrentPriorityQueue.cs
@@ -6,39 +6,38 @@ using CSharpFunctionalExtensions;
 
 namespace ConcurrentPriorityQueue.Core
 {
-    public class ConcurrentPriorityQueue<T, TP> : IConcurrentPriorityQueue<T, TP> 
-	    where T : IHavePriority<TP>
-		where TP : IEquatable<TP>, IComparable<TP>
+    public class ConcurrentPriorityQueue<T, TP> : IConcurrentPriorityQueue<T, TP>
+        where T : IHavePriority<TP>
+        where TP : IEquatable<TP>, IComparable<TP>
     {
         private readonly Dictionary<TP, Queue<T>> _internalQueues;
-        private readonly object _syncRoot = new object();
-	    private readonly int _capacity;
+        private readonly int _capacity;
 
-        public int Count => _internalQueues.Count == 0 ? _internalQueues.Count : 
-	        _internalQueues.Values.Select(q => q.Count).Aggregate((a, b) => a + b);
+        public ConcurrentPriorityQueue(int capacity = 0)
+        {
+            _internalQueues = new Dictionary<TP, Queue<T>>();
+            _capacity = capacity;
+        }
+
+        public int Count => _internalQueues.Count == 0 ? _internalQueues.Count :
+            _internalQueues.Values.Select(q => q.Count).Aggregate((a, b) => a + b);
 
         public bool IsSynchronized => true;
 
-        public object SyncRoot => _syncRoot;
+        public object SyncRoot { get; } = new object();
 
-		public ConcurrentPriorityQueue(int capacity = 0)
-		{
-			_internalQueues = new Dictionary<TP, Queue<T>>();
-			_capacity = capacity;
-		}
-
-		public void CopyTo(T[] array, int index)
+        public void CopyTo(T[] array, int index)
         {
             var itemsArray = ToArray();
-            lock (_syncRoot)
+            lock (SyncRoot)
                 itemsArray.CopyTo(array, index);
         }
 
         public void CopyTo(Array array, int index) => CopyTo((T[])array, index);
-        
+
         public T[] ToArray()
         {
-            lock (_syncRoot)
+            lock (SyncRoot)
             {
                 var itemsList = new List<T>();
                 foreach (var queue in _internalQueues.OrderBy(q => q.Key).Select(q => q.Value))
@@ -53,17 +52,17 @@ namespace ConcurrentPriorityQueue.Core
         public bool TryAdd(T item) => Enqueue(item).IsSuccess;
 
         public bool TryTake(out T item)
-        {           
+        {
             var result = Dequeue();
-            lock (_syncRoot)
+            lock (SyncRoot)
             {
-                item = default(T);
+                item = default;
                 if (result.IsFailure)
                     return false;
 
                 item = result.Value;
                 return true;
-            }           
+            }
         }
 
         public IEnumerator<T> GetEnumerator() => ((IEnumerable<T>)ToArray()).GetEnumerator();
@@ -72,12 +71,12 @@ namespace ConcurrentPriorityQueue.Core
 
         public Result Enqueue(T item)
         {
-			lock (_syncRoot) return AddOrUpdate(item);      
+            lock (SyncRoot) return AddOrUpdate(item);
         }
 
         public Result<T> Dequeue()
         {
-            lock (_syncRoot)
+            lock (SyncRoot)
             {
                 return GetNextQueue()
                     .OnSuccess(q => q.Dequeue())
@@ -87,7 +86,7 @@ namespace ConcurrentPriorityQueue.Core
 
         public Result<T> Peek()
         {
-            lock (_syncRoot)
+            lock (SyncRoot)
             {
                 return GetNextQueue()
                     .OnSuccess(q => q.Peek())
@@ -110,19 +109,19 @@ namespace ConcurrentPriorityQueue.Core
             return Result.Fail<Queue<T>>("Could not find a queue with items.");
         }
 
-	    private Result AddOrUpdate(T item)
-	    {
-		    if (!_internalQueues.ContainsKey(item.Priority))
-		    {
-			    if (IsAtMaxCapacity())
-				    return Result.Fail("Reached max capacity.");
-			    _internalQueues.Add(item.Priority, new Queue<T>());
-			}
+        private Result AddOrUpdate(T item)
+        {
+            if (!_internalQueues.ContainsKey(item.Priority))
+            {
+                if (IsAtMaxCapacity())
+                    return Result.Fail("Reached max capacity.");
+                _internalQueues.Add(item.Priority, new Queue<T>());
+            }
 
-			_internalQueues[item.Priority].Enqueue(item);
-		    return Result.Ok();
-	    }
+            _internalQueues[item.Priority].Enqueue(item);
+            return Result.Ok();
+        }
 
-	    private bool IsAtMaxCapacity() => _capacity != 0 && Count == _capacity;
+        private bool IsAtMaxCapacity() => _capacity != 0 && Count == _capacity;
     }
 }
