+++ ConcurrentPriorityQueue/Core/IConcurrentPriorityQueue.cs
@@ -1,11 +1,11 @@
-﻿using System.Collections.Concurrent;
-using System;
+﻿using System;
+using System.Collections.Concurrent;
 
 namespace ConcurrentPriorityQueue.Core
 {
-	public interface IConcurrentPriorityQueue<T, TP> : IProducerConsumerCollection<T>, IPriorityQueue<T> 
-		where T : IHavePriority<TP>
-		where TP : IEquatable<TP>, IComparable<TP>
-	{
-	}
+    public interface IConcurrentPriorityQueue<T, TP> : IProducerConsumerCollection<T>, IPriorityQueue<T>
+        where T : IHavePriority<TP>
+        where TP : IEquatable<TP>, IComparable<TP>
+    {
+    }
 }
