+++ ConcurrentPriorityQueue/Core/IHavePriority.cs
@@ -2,8 +2,9 @@
 
 namespace ConcurrentPriorityQueue.Core
 {
-	public interface IHavePriority<TP> where TP : IEquatable<TP>, IComparable<TP>
-	{
-		TP Priority { get; set; }
-	}
+    public interface IHavePriority<TP>
+        where TP : IEquatable<TP>, IComparable<TP>
+    {
+        TP Priority { get; set; }
+    }
 }
