+++ ConcurrentPriorityQueue/Core/IPriorityQueue.cs
@@ -2,12 +2,12 @@
 
 namespace ConcurrentPriorityQueue.Core
 {
-	public interface IPriorityQueue<T>
-	{
-		Result Enqueue(T item);
+    public interface IPriorityQueue<T>
+    {
+        Result Enqueue(T item);
 
-		Result<T> Dequeue();
+        Result<T> Dequeue();
 
-		Result<T> Peek();
-	}
+        Result<T> Peek();
+    }
 }
