+++ ConcurrentPriorityQueueTests/MockWithIntegerPriority.cs
@@ -1,15 +1,14 @@
 ﻿using ConcurrentPriorityQueue.Core;
 
-
 namespace ConcurrentPriorityQueueTests
-{	
-	public class MockWithIntegerPriority : IHavePriority<int>
-	{
-		public MockWithIntegerPriority(int priority = 0)
-		{
-			Priority = priority;
-		}
+{
+    public class MockWithIntegerPriority : IHavePriority<int>
+    {
+        public MockWithIntegerPriority(int priority = 0)
+        {
+            Priority = priority;
+        }
 
-		public int Priority { get; set; }		
-	}	
+        public int Priority { get; set; }
+    }
 }
