+++ GenericConcurrentPriorityQueueTests/MockWithObjectPriority.cs
@@ -2,13 +2,13 @@
 
 namespace GenericConcurrentPriorityQueueTests
 {
-	public class MockWithObjectPriority : IHavePriority<TimeToProcess>
-	{
-		public MockWithObjectPriority(TimeToProcess timeToProcess)
-		{
-			Priority = timeToProcess;
-		}
+    public class MockWithObjectPriority : IHavePriority<TimeToProcess>
+    {
+        public MockWithObjectPriority(TimeToProcess timeToProcess)
+        {
+            Priority = timeToProcess;
+        }
 
-		public TimeToProcess Priority { get; set; }
-	}	
+        public TimeToProcess Priority { get; set; }
+    }
 }
