+++ GenericConcurrentPriorityQueueTests/TimeToProcess.cs
@@ -2,21 +2,21 @@
 
 namespace GenericConcurrentPriorityQueueTests
 {
-	public class TimeToProcess : IEquatable<TimeToProcess>, IComparable<TimeToProcess>
-	{
-		public TimeToProcess(decimal timeToProcessInMilliseconds = 0)
-		{
-			TimeInMilliseconds = timeToProcessInMilliseconds;
-		}
+    public class TimeToProcess : IEquatable<TimeToProcess>, IComparable<TimeToProcess>
+    {
+        public TimeToProcess(decimal timeToProcessInMilliseconds = 0)
+        {
+            TimeInMilliseconds = timeToProcessInMilliseconds;
+        }
 
-		public decimal TimeInMilliseconds { get; }
+        public decimal TimeInMilliseconds { get; }
 
-		public int CompareTo(TimeToProcess other) =>
-			TimeInMilliseconds.CompareTo(other.TimeInMilliseconds);
+        public int CompareTo(TimeToProcess other) =>
+            TimeInMilliseconds.CompareTo(other.TimeInMilliseconds);
 
-		public bool Equals(TimeToProcess other) =>
-			TimeInMilliseconds.Equals(other.TimeInMilliseconds);
+        public bool Equals(TimeToProcess other) =>
+            TimeInMilliseconds.Equals(other.TimeInMilliseconds);
 
-		public override int GetHashCode() => TimeInMilliseconds.GetHashCode();
-	}
+        public override int GetHashCode() => TimeInMilliseconds.GetHashCode();
+    }
 }
